+++ StoneOfAdventure_2019_UnityProject/Assets/Scripts/Combat/IDamaged.cs
@@ -1,5 +1,5 @@
 ﻿
 public interface IDamaged
 {
-    void ApplyDamage(float damage);
+    void ApplyDamage(int damage);
 }
